Refactored error estimators into one file
