+++ mcos/error_estimator.py
@@ -15,3 +15,17 @@ class AbstractErrorEstimator(ABC):
         @return
         """
         pass
+
+
+class ExpectedOutcomeErrorEstimator(AbstractErrorEstimator):
+    """Error Estimator that calculates the mean difference in expected outcomes"""
+
+    def estimate(self, mu: np.array, cov: np.array, allocation: np.array, optimal_allocation: np.array) -> float:
+        return np.dot(optimal_allocation - allocation, mu)
+
+
+class VarianceErrorEstimator(AbstractErrorEstimator):
+    """Error Estimator that calculates the mean difference in variance"""
+
+    def estimate(self, mu: np.array, cov: np.array, allocation: np.array, optimal_allocation: np.array) -> float:
+        return np.dot(optimal_allocation - allocation, np.dot(optimal_allocation - allocation, cov))
